modified:   .gitignore 	modified:   CMakeLists.txt 	modified:   lib/mbedtls_config.h 	modified:   src/console.c 	modified:   src/core.h 	modified:   src/crypto.c 	modified:   src/eventloop.c 	modified:   src/ffi.c 	modified:   src/fs.c 	modified:   src/pipe.c 	modified:   src/polyfill.h 	modified:   src/process.c 	modified:   src/repl.h 	modified:   src/repl.js 	modified:   src/socket.c 	modified:   src/thread.c 	modified:   test/hsserver.js 	modified:   test/signal.js 	modified:   types/socket.d.ts 	modified:   types/stdio.d.ts

+++ src/console.c
@@ -491,7 +491,7 @@ __function_break:
         }
 
         if(depth >= MAX_DEPTH || check_circular(ctx, val, visited, depth)){
-            dbuf_printf(output, ANSI_RED "ArrayLike(%lld)" ANSI_RESET, length); // 保留格式化输出
+            dbuf_printf(output, ANSI_RED "ArrayLike(%ld)" ANSI_RESET, (long int)length); // 保留格式化输出
             goto end;
         }
